I'm dumb and the update person endpoint wans't working for pictures, problem solved

+++ BackEnd/Cookly/src/main/java/com/michelfilho/cookly/person/service/PersonService.java
@@ -90,8 +90,10 @@ public class PersonService {
     }
 
     private void updateProfilePicture(Person person, MultipartFile image) {
-        imageService.deleteImage(profilePicturesPath + "/" + person.getProfilePicturePath());
-        imageService.saveImage(profilePicturesPath, image);
+        if(!person.getProfilePictureName().isBlank())
+            imageService.deleteImage(profilePicturesPath + "/" + person.getProfilePicturePath());
+
+        person.setProfilePicturePath(imageService.saveImage(profilePicturesPath, image));
     }
 
 }
